Revert "Merge branch 'develop' into feature/sjy/TG-5"
This reverts commit a14fda89c2f90b8888e414e5efe8e6a931ac68f6, reversing
changes made to 6042779ba61e0971d30fd31f71b056dbeb7df000.
