+++ package.json
@@ -15,14 +15,10 @@
     "@radix-ui/react-toast": "^1.1.5",
     "@tanstack/react-query": "^5.14.6",
     "@tanstack/react-query-devtools": "^5.14.7",
-    "@types/file-saver": "^2.0.7",
     "add": "^2.0.6",
     "axios": "^1.6.2",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.0.0",
-    "file-saver": "^2.0.5",
-    "html-to-image": "^1.11.11",
-    "html2canvas": "^1.4.1",
     "lucide-react": "^0.302.0",
     "next": "14.0.4",
     "next-auth": "^4.24.5",
