+++ src/app/[userId]/page.tsx
@@ -61,7 +61,7 @@ export default async function DishPage({ params: { userId } }: Props) {
           </div>
 
           <PaginationEntire />
-          <ShareButton />
+          <ShareButton btnType="snap-shot" />
         </div>
       </div>
     </section>
