--- src/components/SaveImage.tsx
@@ -1,33 +1,0 @@
-import Image from 'next/image'
-import React from 'react'
-
-const SaveImage = () => {
-  const tempData = {
-    hostAvatar: 'avatar1',
-    visitorAvatar: 'defaultAvatar',
-  }
-  return (
-    <>
-      <div className="flex-center flex flex-row ">
-        <Image
-          src={`/images/avatar/${tempData.hostAvatar}.png`}
-          width={125}
-          height={158}
-          alt="hostAvatar"
-        />
-        <Image
-          src={`/images/avatar/${tempData.visitorAvatar}.png`}
-          width={125}
-          height={158}
-          alt="visitorAvatar"
-        />
-      </div>
-      <div className="flex-center mb-95 flex flex-row gap-13">
-        <div className=" flex-center flex h-108 w-108 rounded-full border bg-gr-500">{'떡국'}</div>
-        <div className=" flex-center flex h-108 w-108 rounded-full border bg-gr-500">{'떡국'}</div>
-      </div>
-    </>
-  )
-}
-
-export default SaveImage
